:recycle:  clean up types
