+++ packages/types/src/logger.d.ts
@@ -1,0 +1,8 @@
+type LogLevel = "debug" | "error" | "info" | "log" | "success" | "warn";
+
+export type LoggerType = {
+  _log: (message: string, level?: LogLevel) => void;
+  instance: {
+    [level in LogLevel]?: (...unknown) => void;
+  };
+};
